Centralise all path utilities in `utils/path.ts`
